Ozuna dev

+++ src/handlers/usersHandler.js
@@ -1,14 +1,27 @@
+import { where } from "sequelize";
 import {
   getAllUsersController,
   getUserByIdController,
   updateUserController,
   deleteUserController,
   registercontroller,
+  registerAuth0controller,
   loginController,
+  updateProfileController,
 } from "../controllers/usersController.js";
+import { Op } from "sequelize";
 import { User } from "../database/db.js";
 
 const getAllUsersHandler = async (req, res) => {
+  // Obtener el rol del usuario autenticado desde la solicitud
+  const userRole = req.user.role;
+  // // Verificar si el usuario autenticado tiene permiso para actualizar
+  if (userRole !== "admin") {
+    return res
+      .status(403)
+      .send({ message: "Unauthorized operation: User is not an admin" });
+  }
+
   try {
     const users = await getAllUsersController();
     res.status(200).send(users);
@@ -20,9 +33,6 @@ const getAllUsersHandler = async (req, res) => {
 const getUserByIdHandler = async (req, res) => {
   try {
     const { id } = req.params;
-    // Obtener el rol del usuario autenticado desde la solicitud
-    const userRole = req.user.role;
-    if(userRole !== 'admin') return res.status(400).json({error:"User Unauthorized"});;
     const user = await getUserByIdController(id);
     if (!user) throw new Error("User not found");
     res.status(200).send(user);
@@ -31,38 +41,55 @@ const getUserByIdHandler = async (req, res) => {
   }
 };
 
-const registerHandler = async (req,res) =>{
-  const {email,password,name} =req.body;
+const genereteAuth0User = async (req, res) => {
+  const { email, name } = req.body;
   try {
-    const user = await registercontroller(email,password,name);
+    const user = await registerAuth0controller(email, name);
     res.status(200).json(user);
-  }catch(error){
-    res.status(400).json({message:error.message})
-  }
-}
-
-const loginHandler = async (req,res) =>{
-  const {email,password} = req.body;
-  try {
-    const userExisting = await User.findOne({where: {email}});
-    if(!userExisting){ 
-      return res.status(400).json({error:"Email not found"});
-    }
-    if(!userExisting.isActive){ 
-      return res.status(400).json({error:"User not Active"});
-    }
-    const token = await loginController(userExisting,password);
-    res.cookie("token",token,{httpOnly:true});
-    res.json({ message: "Inicio de sesión exitoso", token });
-    // res.json(authenticatedUser);
-    // console.log("Ingreso exitoso");
-    
   } catch (error) {
-    res.status(400).json({message:error.message});    
-
+    res.status(400).json({ message: error.message });
   }
-}
+};
 
+const registerHandler = async (req, res) => {
+  const { email, password, name } = req.body;
+const userExisting = await User.findOne({where: {email}});
+if (userExisting)  return res.status(400).json({ message: "User already exists" });
+  try {
+    const user = await registercontroller(email, password, name);
+    res.status(200).json(user);
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+};
+
+const loginHandler = async (req, res) => {
+  const { email, password, name } = req.body;
+  console.log(email,password,name);
+  let userExisting = null;
+  try {
+    if (email != undefined) {
+      userExisting = await User.findOne({
+        where: { email: email },
+      });
+    } else {
+      userExisting = await User.findOne({
+        where: { name: name },
+      });
+    }
+    if (!userExisting) {
+      return res.status(400).json({ error: "Email not found" });
+    }
+    if (!userExisting.isActive) {
+      return res.status(400).json({ error: "User not Active" });
+    }
+    const token = await loginController(userExisting, password);
+    res.cookie("token", token, { httpOnly: true });
+    res.json({ message: "Inicio de sesión exitoso", token });
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+};
 
 const updateUserHandler = async (req, res) => {
   try {
@@ -71,7 +98,9 @@ const updateUserHandler = async (req, res) => {
     const userRole = req.user.role;
     // // Verificar si el usuario autenticado tiene permiso para actualizar
     if (userRole !== "admin") {
-        return res.status(403).send({ message: "Unauthorized operation: User is not an admin" });
+      return res
+        .status(403)
+        .send({ message: "Unauthorized operation: User is not an admin" });
     }
 
     // Obtener los campos actualizados del cuerpo de la solicitud
@@ -85,16 +114,33 @@ const updateUserHandler = async (req, res) => {
   }
 };
 
+const updateProfileHandler = async (req, res) => {
+  try {
+    const { email, id } = req.params;
+    const updatedFields = req.body;
+
+    const updatedUser = await updateProfileController(
+      id ? id : email,
+      updatedFields
+    );
+    res.status(200).send(updatedUser);
+  } catch (error) {
+    res.status(500).send({ message: error.message });
+  }
+};
+
 const deleteUserHandler = async (req, res) => {
   try {
     const { id } = req.params;
     // Obtener el rol del usuario autenticado desde la solicitud
-    //  const userRole = req.user.role;
+    const userRole = req.user.role;
+    // // Verificar si el usuario autenticado tiene permiso para actualizar
+    if (userRole !== "admin") {
+      return res
+        .status(403)
+        .send({ message: "Unauthorized operation: User is not an admin" });
+    }
 
-    //  // Verificar si el usuario autenticado tiene permiso para actualizar
-    //  if (userRole !== "admin") {
-    //      return res.status(403).send({ message: "Unauthorized operation: User is not an admin" });
-    //  }
     const deletedUser = await deleteUserController(id);
     res.status(200).send(deletedUser);
   } catch (error) {
@@ -108,5 +154,7 @@ export {
   updateUserHandler,
   deleteUserHandler,
   registerHandler,
+  genereteAuth0User,
   loginHandler,
+  updateProfileHandler,
 };
